Refactor: New service for userPage that makes use of the repository and interface
Due to the mess that was using two constructors in one service I decided to create a new Service specific for UserPage.
