+++ HealthCareABApi/HealthCareABApi.csproj
@@ -31,6 +31,5 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Configurations\" />
-    <Folder Include="Services\" />
   </ItemGroup>
 </Project>
